% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{classification.accuracy}
\alias{classification.accuracy}
\title{Calculate the Classification Accuracy of a Simple Confusion Matrix}
\usage{
classification.accuracy(conf_matrix)
}
\arguments{
\item{conf_matrix}{The confusion matrix object {TN, FN, FP, TP}}
}
\value{
A number representing the accuracy percentage
}
\description{
This function takes one argument, the confusion matrix, and
calculates the classification accuracy of the matrix by adding
the true negative (index 1) and true positive (index 4) values
and the dividing the result by the sum of the entire matrix.
}
