% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit.R
\docType{data}
\name{credit}
\alias{credit}
\title{Credit card application approval or disapproval}
\format{A data frame with 690 observations and 16 variables
\describe{
  \item{A1}{b, a}
  \item{A2}{continuous}
  \item{A3}{continuous}
  \item{A4}{u, y, l, t}
  \item{A5}{g, p, gg}
  \item{A6}{c, d, cc, i, j, k, m, r, q, w, x, e, aa, ff}
  \item{A7}{v, h, bb, j, n, z, dd, ff, o}
  \item{A8}{continuous}
  \item{A9}{t, f}
  \item{A10}{t, f}
  \item{A11}{continuous}
  \item{A12}{t, f}
  \item{A13}{g, p, s}
  \item{A14}{continuous}
  \item{A15}{continuous}
  \item{class}{+, -}
}}
\source{
\href{http://archive.ics.uci.edu/ml/datasets/Credit+Approval}{Credit Approval Data Set}
}
\usage{
data(credit)
}
\description{
This dataset was sourced from the UCI Machine Learning Repository
\url{http://archive.ics.uci.edu/ml/datasets/Credit+Approval}. It contains
meaningless variable names and values to protect confidentiality.
}
\keyword{datasets}
