% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbst.R
\docType{package}
\name{dbst}
\alias{dbst}
\alias{dbst-package}
\title{dbst: A package collecting functions and data sets used for DBST 667}
\description{
The dbst package contains every data set used throughout the course:
  credit - The CreditApproval dataset.
  cars   - The Cars dataset.
}
